working on solution to swift 5.5.2 async protocol bug.
