+++ AppSyncRealTimeClient/ConnectionProvider/AppsyncRealtimeConnection/RealtimeConnectionProviderAsync+MessageInterceptableAsync.swift
@@ -7,14 +7,14 @@
 
 import Foundation
 
-//@available(iOS 13.0.0, *)
-extension RealtimeConnectionProvider: MessageInterceptableAsync {
+@available(iOS 13.0.0, *)
+extension RealtimeConnectionProviderBase: MessageInterceptableAsync {
     public func addInterceptor(_ interceptor: MessageInterceptorAsync) {
         useAsyncInterceptors = true
         messageInterceptors.append(interceptor)
     }
 
-    func interceptMessage(_ message: AppSyncMessage, for endpoint: URL) async -> AppSyncMessage {
+    public func interceptMessage(_ message: AppSyncMessage, for endpoint: URL) async -> AppSyncMessage {
         guard let messageInterceptors = messageInterceptors as? [MessageInterceptorAsync] else {
             AppSyncLogger.error("Failed to cast messageInterceptors.")
             return message
