+++ AppSyncRealTimeClient/ConnectionProvider/InterceptableConnection.swift
@@ -19,7 +19,7 @@ public protocol ConnectionInterceptable {
 }
 
 /// Intercepts the connect request
-//@available(iOS 13.0.0, *)
+@available(iOS 13.0.0, *)
 public protocol ConnectionInterceptableAsync {
 
     /// Add a new interceptor to the object.
@@ -37,7 +37,7 @@ public protocol MessageInterceptable {
     func interceptMessage(_ message: AppSyncMessage, for endpoint: URL) -> AppSyncMessage
 }
 
-//@available(iOS 13.0.0, *)
+@available(iOS 13.0.0, *)
 public protocol MessageInterceptableAsync {
 
     func addInterceptor(_ interceptor: MessageInterceptorAsync)
@@ -50,7 +50,7 @@ public protocol ConnectionInterceptor {
     func interceptConnection(_ request: AppSyncConnectionRequest, for endpoint: URL) -> AppSyncConnectionRequest
 }
 
-//@available(iOS 13.0.0, *)
+@available(iOS 13.0.0, *)
 public protocol ConnectionInterceptorAsync {
 
     func interceptConnection(_ request: AppSyncConnectionRequest, for endpoint: URL) async -> AppSyncConnectionRequest
@@ -61,7 +61,7 @@ public protocol MessageInterceptor {
     func interceptMessage(_ message: AppSyncMessage, for endpoint: URL) -> AppSyncMessage
 }
 
-//@available(iOS 13.0.0, *)
+@available(iOS 13.0.0, *)
 public protocol MessageInterceptorAsync {
 
     func interceptMessage(_ message: AppSyncMessage, for endpoint: URL) async -> AppSyncMessage
@@ -69,5 +69,5 @@ public protocol MessageInterceptorAsync {
 
 public protocol AuthInterceptor: MessageInterceptor, ConnectionInterceptor {}
 
-//@available(iOS 13.0.0, *)
+@available(iOS 13.0.0, *)
 public protocol AuthInterceptorAsync: MessageInterceptorAsync, ConnectionInterceptorAsync {}
